+++ database/migrations/2023_07_12_000000_add_asset_data_field_and_type.php
@@ -1,21 +1,23 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up()
     {
-        if(Schema::hasColumn('assets','data')) return;
+        if (Schema::hasColumn('assets', 'data')) {
+            return;
+        }
 
         Schema::table('assets', function (Blueprint $table) {
             $table->string('asset_type')->nullable();
             $table->json('data')->nullable();
         });
 
-        \Illuminate\Support\Facades\DB::table('assets')->update(['asset_type' => 'local']);
+        DB::table('assets')->update(['asset_type' => 'default']);
 
         Schema::table('assets', function (Blueprint $table) {
             $table->string('asset_type')->nullable(false)->change(); // Remove nullable